requirements.txt
click
fastapi
h11
joblib
numpy
pandas
pydantic
python-dateutil
python-multipart
pytz
six
starlette
threadpoolctl
uvicorn
Pillow
tritonclient[all]
tqdm
psycopg2